requirements.txt
-i https://pypi.org/simple
aiofile==3.8.8; python_version >= '3.7' and python_version < '4'
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiosignal==1.3.1; python_version >= '3.7'
asyncclick==8.1.3.4; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
bidict==0.22.1; python_version >= '3.7'
brotli==1.1.0
caio==0.9.13; python_version >= '3.7' and python_version < '4'
cffi==1.16.0; python_version >= '3.8'
colorama==0.4.6; platform_system == 'Windows'
cssselect2==0.7.0; python_version >= '3.7'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
ensure==1.0.3; python_version >= '3.6'
fonttools==4.46.0; python_version >= '3.8'
fpdf==1.7.2
fpdf2==2.7.6; python_version >= '3.7'
frozenlist==1.4.0; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.40
h11==0.14.0; python_version >= '3.7'
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.6; python_version >= '3.5'
lxml==4.9.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markdown2==2.4.11; python_version >= '3.5' and python_version < '4'
mpmath==1.3.0
multidict==6.0.4; python_version >= '3.7'
pillow==10.1.0; python_version >= '3.8'
pycparser==2.21
pydyf==0.8.0; python_version >= '3.7'
pyphen==0.14.0; python_version >= '3.7'
python-engineio==4.8.0; python_version >= '3.6'
python-socketio==5.10.0
setuptools==69.0.2
simple-websocket==1.0.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
svgmanip==0.0.12
svgutils==0.3.4
tabulate==0.9.0; python_version >= '3.7'
file:////Users/oliverbarnum/Documents/git/templative-electron/dependencies/templative-0.0.1.tar.gz
tinycss2==1.2.1; python_version >= '3.7'
wand==0.6.13
weasyprint==60.1; python_version >= '3.7'
webencodings==0.5.1
wsproto==1.2.0; python_full_version >= '3.7.0'
yarl==1.9.4; python_version >= '3.7'
zopfli==0.2.3
